pyproject.toml
[tool.poetry]
name = "checkbox"
version = "0.1.0"
description = ""
authors = ["Andrii <tap.press.swap@gmail.com>"]
packages = [
    { include = "checkbox", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.0"}
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
python-ulid = "^2.7.0"
sqlalchemy-utils = "^0.41.2"
dishka = "^1.3.0"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
pytest-asyncio = "^0.24.0"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
